feat(api): Office and PDF profiles — canonicalizing ooxml-v1, byte-exact ooxml-ber-v1, pdf-v1 (#15)
- ooxml-v1 (id 8): canonicalizing Office profile — members inflated (stored/deflate) and re-emitted deterministically (STORE, zeroed metadata, original order/names); chunks concatenate to the canonical form; part-aligned cuts at the 4 KiB document-unit minimum; CRC/size verification against source claims; signed packages fail hard toward BER; PackageObserver event tap reports names, canonical offsets, and inflated member bytes

- ooxml-ber-v1 (id 9): the same part-aware walk over the original bytes, byte-exact, signatures survive; policy-selected, never probed or registry-mapped

- pdf-v1 (id 10): object-aligned boundaries — header, indirect objects (direct /Length skip with deterministic scan fallback, indirect via endstream scan), xref tables, trailers, incremental updates

- ZIP walker gains a ZipEvents tap (zip-v1 unchanged); BoundaryAssembler gains a unit-min parameter; ProfileSet widens to u16

- ProfileRegistry::V2 + Detector::V2 (defaults): OOXML types move zip->ooxml, application/pdf added; V1 stays frozen

- README: profiles table rows, canonical-vs-BER section, reuse walkthrough updates; version 0.3.0

+++ src/probe/set.rs
@@ -7,7 +7,7 @@ use crate::profile::ChunkingProfile;
 /// A set of registry profiles: one bit per entry of
 /// [`ChunkingProfile::ALL`].
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Default)]
-pub struct ProfileSet(u8);
+pub struct ProfileSet(u16);
 
 impl ProfileSet {
     /// The empty set.
@@ -21,7 +21,7 @@ impl ProfileSet {
         set
     }
 
-    fn bit(profile: ChunkingProfile) -> u8 {
+    fn bit(profile: ChunkingProfile) -> u16 {
         let index = ChunkingProfile::ALL
             .iter()
             .position(|entry| *entry == profile)
